chore: moves api-specs to tests
