+++ .detoxrc.js
@@ -41,7 +41,7 @@ module.exports = {
       app: process.env.CI ? `ios.${process.env.METAMASK_BUILD_TYPE}.release` : 'ios.debug',
       testRunner: {
         args: {
-          "$0": "node e2e/api-specs/run-api-spec-tests.js",
+          "$0": "node tests/smoke/api-specs/run-api-spec-tests.js",
         },
       },
     },
